requirements.txt
streamlit
python-dotenv
plotly
passlib
streamlit-chat
geopy
reportlab
streamlit-folium